Change configs to use Doubles

+++ src/machines/java/com/enderio/machines/common/config/client/MachinesClientConfig.java
@@ -6,13 +6,13 @@ public class MachinesClientConfig {
 
     public final MachinesClientBlocksConfig BLOCKS;
 
-    public final ForgeConfigSpec.ConfigValue<Float> IO_CONFIG_NEIGHBOUR_TRANSPARENCY;
+    public final ForgeConfigSpec.ConfigValue<Double> IO_CONFIG_NEIGHBOUR_TRANSPARENCY;
 
     public MachinesClientConfig(ForgeConfigSpec.Builder builder) {
         BLOCKS = new MachinesClientBlocksConfig(builder);
 
         builder.push("ioconfig");
-        IO_CONFIG_NEIGHBOUR_TRANSPARENCY = builder.comment("Neighbour Transparency [0-1]").define("neighbourTransparency", 0.4F);
+        IO_CONFIG_NEIGHBOUR_TRANSPARENCY = builder.comment("Neighbour Transparency [0-1]").define("neighbourTransparency", 0.4d);
         builder.pop();
     }
 }
