+++ src/machines/java/com/enderio/machines/common/config/common/MachinesCommonConfig.java
@@ -5,8 +5,8 @@ import net.minecraftforge.common.ForgeConfigSpec;
 
 public class MachinesCommonConfig {
     public final EnergyConfig ENERGY;
-    public final ForgeConfigSpec.ConfigValue<Float> ENCHANTER_LAPIS_COST_FACTOR;
-    public final ForgeConfigSpec.ConfigValue<Float> ENCHANTER_LEVEL_COST_FACTOR;
+    public final ForgeConfigSpec.ConfigValue<Double> ENCHANTER_LAPIS_COST_FACTOR;
+    public final ForgeConfigSpec.ConfigValue<Double> ENCHANTER_LEVEL_COST_FACTOR;
     public final ForgeConfigSpec.ConfigValue<Integer> ENCHANTER_BASE_LEVEL_COST;
     public final ForgeConfigSpec.ConfigValue<Integer> MAX_SPAWNER_ENTITIES;
     public final ForgeConfigSpec.ConfigValue<Integer> MAX_SPAWNERS;
@@ -17,8 +17,8 @@ public class MachinesCommonConfig {
         ENERGY = new EnergyConfig(builder);
 
         builder.push("enchanter");
-            ENCHANTER_LAPIS_COST_FACTOR = builder.comment("The lapis cost is enchant level multiplied by this value.").define("lapisCostFactor", 3.0f);
-            ENCHANTER_LEVEL_COST_FACTOR = builder.comment("The final XP cost for an enchantment is multiplied by this value. To halve costs set to 0.5, to double them set it to 2.").define("levelCostFactor", 0.75f);
+            ENCHANTER_LAPIS_COST_FACTOR = builder.comment("The lapis cost is enchant level multiplied by this value.").define("lapisCostFactor", 3.0d);
+            ENCHANTER_LEVEL_COST_FACTOR = builder.comment("The final XP cost for an enchantment is multiplied by this value. To halve costs set to 0.5, to double them set it to 2.").define("levelCostFactor", 0.75d);
             ENCHANTER_BASE_LEVEL_COST = builder.comment("Base level cost added to all recipes in the enchanter.").define("baseLevelCost", 2);
         builder.pop();
